Make it OO, and generalize half-edge data structure consutrction.
First major thing is that the program is now modular and tries to
go with an OO design. There's still some more things that can be
done to improve this though.

In addition, n-sided faces for n > 3 are parsed properly and
are properly built into a half-edge mesh data structure.
Further, n-sided faces can now be properly oriented.

As a result of the above, triangulation of the .obj file before
construction is not necessary anymore.